requirements.txt

fastapi
uvicorn
chromadb>=0.5.0
openai>=1.0.0
pydantic
tiktoken
python-dotenv
